+++ DerailValleyUniversalJato/InGameWindow.cs
@@ -3,17 +3,13 @@ using UnityEngine;
 using UnityModManagerNet;
 using DerailValleyModToolbar;
 using DV;
+using DV.ThingTypes;
 
 namespace DerailValleyUniversalJato;
 
 public class InGameWindow : MonoBehaviour, IModToolbarPanel
 {
     private UnityModManager.ModEntry.ModLogger Logger => Main.ModEntry.Logger;
-    private bool showGui = false;
-    private Rect buttonRect = new Rect(60, 30, 20, 20); // TODO: avoid conflict with other mods (currently just DV Utilities mod)
-    private Rect windowRect = new Rect(20, 30, 0, 0);
-    private Rect scrollRect;
-    private Vector2 scrollPosition;
     public static JatoSettings NewSettings = new JatoSettings();
     private string _keyCodeText = NewSettings.KeyCode.ToString();
     private string _thrustText = NewSettings.Thrust.ToString();
@@ -24,7 +20,6 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
     private string _rotationYText = NewSettings.RotationY.ToString();
     private string _rotationZText = NewSettings.RotationZ.ToString();
     private string _scaleText = NewSettings.Scale.ToString();
-    private string _volumeText = NewSettings.SoundVolume.ToString();
     private int? _selectedComponentIndex = null;
 
     (Transform transform, Rigidbody Rigidbody, TrainCar trainCar)? GetJatoTargetInfo()
@@ -38,19 +33,19 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
     void AddJato()
     {
         var target = GetJatoTargetInfo();
-
         if (target == null)
             return;
 
         var (transform, rigidbody, trainCar) = target.Value;
 
         JatoManager.AddJato(transform, trainCar, rigidbody, NewSettings);
+
+        _selectedComponentIndex = null;
     }
 
     void AddMirroredJato()
     {
         var target = GetJatoTargetInfo();
-
         if (target == null)
             return;
 
@@ -65,12 +60,13 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
         JatoManager.AddJato(transform, trainCar, rigidbody, NewSettings);
 
         NewSettings.PositionX = oldPosX;
+
+        _selectedComponentIndex = null;
     }
 
     void UpdateJato(bool applyOffsets = true)
     {
         var target = GetJatoTargetInfo();
-
         if (target == null)
             return;
 
@@ -82,23 +78,27 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
     void RemoveJato()
     {
         var target = GetJatoTargetInfo();
-
         if (target == null)
             return;
 
         var (transform, rigidbody, trainCar) = target.Value;
 
         JatoManager.RemoveJato(transform, _selectedComponentIndex);
+
+        _selectedComponentIndex = null;
     }
 
     void StopTrainMoving()
     {
         var target = GetJatoTargetInfo();
-
         if (target == null)
             return;
 
         var (transform, rigidbody, trainCar) = target.Value;
+
+        // fix null error from the game
+        if (trainCar == null)
+            return;
 
         Logger.Log("Stopping train moving");
 
@@ -108,7 +108,6 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
     void DerailTrain()
     {
         var target = GetJatoTargetInfo();
-
         if (target == null)
             return;
 
@@ -129,11 +128,97 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
         Globals.G.GameParams.DerailStressThreshold = Globals.G.GameParams.defaultStressThreshold;
     }
 
+    private void UpdateDebugTexts()
+    {
+        var target = GetJatoTargetInfo();
+        var comps = JatoManager.allJatos;
+
+        if (target == null || _selectedComponentIndex == null)
+        {
+            foreach (var comp in comps)
+                comp.DebugText = null;
+        }
+
+        var compsOnTarget = JatoManager.GetJatos(target!.Value.transform);
+
+        for (var i = 0; i < compsOnTarget.Length; i++)
+        {
+            var comp = compsOnTarget[i];
+
+            comp.DebugText = i == _selectedComponentIndex ? $"{i + 1}" : null;
+        }
+    }
+
+    private void AutoPosition()
+    {
+        var target = GetJatoTargetInfo();
+        if (target == null)
+            return;
+
+        var (transform, rigidbody, trainCar) = target.Value;
+
+        var positionX = 0f;
+        var positionY = 0f;
+        var positionZ = 0f;
+
+        switch (trainCar.carType)
+        {
+            case TrainCarType.LocoShunter:
+                positionX = 1.125f;
+                positionY = 1.51f;
+                positionZ = -3f;
+                break;
+                // TODO
+        }
+
+        _positionXText = positionX.ToString();
+        _positionYText = positionY.ToString();
+        _positionZText = positionZ.ToString();
+
+        NewSettings.PositionX = positionX;
+        NewSettings.PositionY = positionY;
+        NewSettings.PositionZ = positionZ;
+    }
+
+    private bool _isRecordingKey = false;
+
     public void Window(Rect rect)
     {
         var target = GetJatoTargetInfo();
 
+        bool? alreadyHasJato = target != null ? JatoManager.GetDoesTargetHaveJato(target.Value.transform) : null;
+
         GUILayout.Label($"Target: {(target != null ? $"{target.Value.trainCar.carType} {target.Value.trainCar.carLivery.id}" : "(none)")}");
+
+        if (target != null)
+        {
+            var count = JatoManager.GetJatoCount(target.Value.transform);
+
+            if (count > 1)
+            {
+                for (var i = 0; i < count; i++)
+                {
+                    if (GUILayout.Button($"#{i + 1}{(i == _selectedComponentIndex ? " ✓" : "")}", GUILayout.Width(50f)))
+                    {
+                        if (_selectedComponentIndex == i)
+                            _selectedComponentIndex = null;
+                        else
+                            _selectedComponentIndex = i;
+                    }
+                }
+            }
+            else
+            {
+                _selectedComponentIndex = null;
+            }
+        }
+        else
+        {
+
+            _selectedComponentIndex = null;
+        }
+
+        UpdateDebugTexts();
 
         GUILayout.Label($"Key:");
         _keyCodeText = GUILayout.TextField(_keyCodeText);
@@ -144,16 +229,33 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
         }
         GUILayout.Label($"(any Unity keycode eg. 'LeftShift', 'E', 'Enter', 'F12', 'UpArrow')");
 
-        GUILayout.Label($"Thrust:");
-        _thrustText = GUILayout.TextField(_thrustText);
+        if (GUILayout.Button("Record Key"))
+        {
+            _isRecordingKey = true;
+        }
 
+        if (_isRecordingKey)
+        {
+            GUILayout.Label("Waiting for you to press a key...");
+
+            var e = Event.current;
+            if (e.type == EventType.KeyDown)
+            {
+                NewSettings.KeyCode = e.keyCode;
+                _keyCodeText = e.keyCode.ToString();
+                _isRecordingKey = false;
+            }
+        }
+
+        GUILayout.Label($"Thrust (newtons):");
+        _thrustText = GUILayout.TextField(_thrustText);
         if (float.TryParse(_thrustText, out float thrustResult))
         {
             NewSettings.Thrust = thrustResult;
         }
 
+        GUILayout.BeginHorizontal();
         string[] forces = ["50,000", "100,000", "250,000", "500,000", "1,000,000"];
-
         foreach (var force in forces)
         {
             if (GUILayout.Button(force))
@@ -161,6 +263,21 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
                 _thrustText = force;
             }
         }
+        GUILayout.EndHorizontal();
+
+        NewSettings.ForceOn = GUILayout.Toggle(NewSettings.ForceOn, "Force on");
+        NewSettings.RequireSittingInside = GUILayout.Toggle(NewSettings.RequireSittingInside, "Must be on train to use");
+        NewSettings.HideBody = GUILayout.Toggle(NewSettings.HideBody, "Only show flame");
+
+        GUILayout.Label($"Volume:");
+        NewSettings.SoundVolume = GUILayout.HorizontalSlider(NewSettings.SoundVolume, 0f, 1f);
+
+        GUI.enabled = alreadyHasJato == true;
+        if (GUILayout.Button(_selectedComponentIndex != null ? $"Update #{_selectedComponentIndex + 1}" : "Update All"))
+        {
+            UpdateJato(applyOffsets: false);
+        }
+        GUI.enabled = true;
 
         GUILayout.BeginHorizontal();
         GUILayout.Label($"Position:");
@@ -181,6 +298,11 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
         if (float.TryParse(_positionZText, out float positionZResult))
         {
             NewSettings.PositionZ = positionZResult;
+        }
+
+        if (GUILayout.Button("Auto", GUILayout.Width(50f)))
+        {
+            AutoPosition();
         }
         GUILayout.EndHorizontal();
 
@@ -204,54 +326,15 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
         {
             NewSettings.RotationZ = rotationZResult;
         }
+
+        GUILayout.Label("", GUILayout.Width(50f));
         GUILayout.EndHorizontal();
 
         GUILayout.Label($"Scale:");
-
         _scaleText = GUILayout.TextField(_scaleText, GUILayout.Width(40f));
         if (float.TryParse(_scaleText, out float scaleResult))
         {
             NewSettings.Scale = scaleResult;
-        }
-
-        NewSettings.ForceOn = GUILayout.Toggle(NewSettings.ForceOn, "Force on");
-        NewSettings.RequireSittingInside = GUILayout.Toggle(NewSettings.RequireSittingInside, "Must sit inside to use");
-
-        GUILayout.Label($"Volume (eg. 0.5 for 50%):");
-        _volumeText = GUILayout.TextField(_volumeText, GUILayout.Width(50f));
-        if (float.TryParse(_rotationZText, out float volumeResult))
-        {
-            NewSettings.SoundVolume = volumeResult;
-        }
-
-        bool? alreadyHasJato = target != null ? JatoManager.GetDoesTargetHaveJato(target.Value.transform) : null;
-
-        if (target != null)
-        {
-            var count = JatoManager.GetJatoCount(target.Value.transform);
-
-            if (count > 1)
-            {
-                for (var i = 0; i < count; i++)
-                {
-                    if (GUILayout.Button($"Rocket #{i + 1}{(i == _selectedComponentIndex ? " - selected" : "")}"))
-                    {
-                        if (_selectedComponentIndex == i)
-                            _selectedComponentIndex = null;
-                        else
-                            _selectedComponentIndex = i;
-                    }
-                }
-            }
-            else
-            {
-                _selectedComponentIndex = null;
-            }
-        }
-        else
-        {
-
-            _selectedComponentIndex = null;
         }
 
         if (GUILayout.Button("Add"))
@@ -259,23 +342,17 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
             AddJato();
         }
         GUI.enabled = alreadyHasJato == true;
-        if (GUILayout.Button("Update Without Offsets"))
-        {
-            UpdateJato(applyOffsets: false);
-        }
-        if (GUILayout.Button("Update Everything"))
+        if (GUILayout.Button(_selectedComponentIndex != null ? $"Replace #{_selectedComponentIndex + 1}" : "Replace All"))
         {
             UpdateJato();
         }
-        if (GUILayout.Button("Remove"))
+        if (GUILayout.Button(_selectedComponentIndex != null ? $"Remove #{_selectedComponentIndex + 1}" : "Remove All"))
         {
             RemoveJato();
         }
         GUI.enabled = true;
 
-        GUILayout.Label($"");
-
-        GUILayout.Label("Make position X greater than 0 and you can add two JATOs one on either side:");
+        GUILayout.Label("If position X is greater than 0:");
         GUI.enabled = NewSettings.PositionX != 0;
         if (GUILayout.Button("Add 2 Mirrored"))
         {
@@ -292,8 +369,6 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
         else
             DisableNoDerail();
 
-        GUILayout.Label($"");
-
         if (GUILayout.Button("Stop Train Moving"))
         {
             StopTrainMoving();
@@ -302,7 +377,7 @@ public class InGameWindow : MonoBehaviour, IModToolbarPanel
         {
             DerailTrain();
         }
-        if (GUILayout.Button("Sync Into Settings"))
+        if (GUILayout.Button("Save Settings"))
         {
             Main.settings.LastJatoSettings = NewSettings.Clone();
         }